refactor : create small change in ItemRepo interface

+++ Backend/springboot-POS-Backend/src/main/java/com/example/springbootposbackend/repository/ItemRepo.java
@@ -3,5 +3,7 @@ package com.example.springbootposbackend.repository;
 import com.example.springbootposbackend.entity.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemRepo extends JpaRepository<Item,Integer> {
+import java.util.List;
+
+public interface ItemRepo extends JpaRepository<Item,Long> {
 }
